requirements.txt
Django==4.2.7
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
face-recognition==1.3.0
Pillow==10.0.1
numpy==1.24.3
dlib==19.24.2
python-decouple==3.8
requests==2.31.0
channels==4.0.0
channels-redis==4.1.0
django-cors-headers==4.3.1
